requirements.txt
dbt-core==1.8.6
dbt-duckdb==1.8.2
duckdb==1.1.3
pandas==2.2.2
requests==2.32.3
python-dateutil==2.9.0.post0
pyyaml==6.0.2
plotly==5.24.1  # not required for CI; dashboard uses CDN
